C++ implementation of the E and M steps from Winkler's EM algorithm estimating FS method using sparse matrices for big sample sizes

C++ implementation of the E and M steps from Winkler's EM algorithm estimating FS method using sparse matrices for big sample sizes

Usage

EMstep_C_sparse_big(mat_A, mat_B, p, m, u)