C++ implementation of the pseudo-likelihood computation

loglikC_bin implements an even faster C++ implementation of the pseudo-likelihood computation for binary variables

loglikC_bin_wDates implements a C++ implementation of the pseudo-likelihood computation for binary variables with dates

Usage

loglikC_bin(Bmat, Amat, eps_p, eps_n, piA, piB)

loglikC_bin_wDates(Bmat, Amat, Bdates, Adates, eps_p, eps_n, piA, piB)

loglikratioC_diff_arbitrary(Bmat, Amat, d_max, cost)

Arguments

Bmat

K x nB matrix of the observations to be matched.

Amat

nA x K matrix the database into which a match is looked for.

eps_p

a vector of length K giving the prior discrepancy rate expected from A to B for the positives, for each variable.

eps_n

a vector of length K giving the prior discrepancy rate expected from A to B for the negatives, for each variable.

piA

a vector of length K giving the prior probabilities of observing each variable in A.

piB

a vector of length K giving the prior probabilities of observing each variable in B.

Bdates

nB x K matrix of the dates for each observations to be matched.

Adates

nA x K matrix of the dates for database into which a match is looked for.

d_max

a numeric vector of length K giving the minimum difference from which it is considered a discrepancy.

cost

a numeric vector of length K giving the arbitrary cost of discrepancy.