Compute the matching probabilities for each pair of observations

C++ version: for each observations in (1:n), all the matching probabilities are computed for the p possible pairs.

Usage

matchProbs_rank_full_C(computed_dist, prop_match)

Arguments

computed_dist

a n x p matrix of computed distances used for ranking.

prop_match

a priori proportion of matches ("rho_1")

Value

a n x p matrix containing the matching probabilities for each pair