Fast C++ computation of the final posterior probabilities in the E-M Winkler's method

matchingScore_C_sparse_big implements a version using sparse matrices. It has a better management of memory but is a little bit slower (indicated for big matrices)

Usage

matchingScore_C(agreemat, m, u, nA, nB)

matchingScore_C_sparse_big(mat_A, mat_B, m, u)

Arguments

agreemat

binary sparse matrix of dimensions N x K containing the agreement rows for each pair of potential matches.

m

vector of length K containing the agreement weights.

u

vector of length K containing the disagreement weights.

nA

integer indicating the number of observations to be matched.

nB

integer indicating the number of observations to be matched with.

mat_A

a nB x K matrix of the observations to be matched.

mat_B

a nA x K matrix of the database into which a match is looked for.