Computes a matching score from agreement vectors and weights

Computes a matching score from agreement vectors and weights

Usage

matching_score(agree, m, u)

Examples

estep_vect <- function(ag_score, p, m, u){
  a <-exp(log(p) + ag_score%*%log(m) + (1-ag_score)%*%log(1-m))
  b <- exp(log(1-p) + ag_score%*%log(u) + (1-ag_score)%*%log(1-u))
  return(cbind(a/(a+b), b/(a+b)))
}